Add support for per-suffix parsers
Override 'read' method in SphinxStandaloneReader to consider the suffix
of the file being processed and potentially choose to use a different
parser. Add 'parsers' argument to constructor taking a dictionary of
suffixes to names of parsers to use for the suffix, as provided by a
configuration variable of the same name.

This adds the ability to do something like:

    parsers = {'.md': 'markdown.parsers'}

...to use markdown.parsers.Parser to parse .md files. Note that the
custom parser is ONLY used to parse files matching the suffix; otherwise
the default (reST) parser is used.