Fix: #1815 Emit warns and use quiet setting to suppress info
Call warn function instead of info when a warning event occurs and
use the app.quiet flag to only call the info function when there is
additional data that can be disabled and the user has not requested it
be suppressed.

Since the Application object warn() method called by the Builder objects
decides whether to simply log the warning or raise an exception if the
user has enabled 'warniserror', it's necessary to ensure any warning
events always use the warn() method.

This inverts the current logic of always calling info() for warnings
and only calling warn() if the `app.quiet` flag is true and removes an
unnecessary info call where it is outputting identical information to
the corresponding call to warn().