Make sure extensions specified with -D override extensions in config file.
Previously, a comma separated list of extensions was read and parsed
correctly. However, the value was then stored in config hashtable,
which was overwritten when the config file was parsed.

Now, the config file is parsed first, and the the entry in the config
hashtable is updated with the overriding extension, if it exists.

Signed-off-by: mulhern <amulhern@redhat.com>