Introduce FileAvoidWrite class
Sphinx relies on file mtimes to determine whether operations such as
source reading need to occur.

When tools like sphinx-apidoc run, they blindly write files, updating the
mtime of files, even if the final file content does not change.

In this patch, we introduce a FileAvoidWrite class that can be used to
only write files if content has changed. This opens the door to less
cache invalidation via mtime changes and thus better performance. This
will be realized in a subsequent commit.