This falls back to the node's astext() method during message extraction.
Using recommonmark/Markdown support,
I was unable to use the existing i18n tooling because it is depending on `rawsource`.
This makes it fall back to using astext() for nodes without a `rawsource`.

This makes it so I am able to translate messages with the i18n support with Commonmark.