Fix broken TOC of PDFs if section includes an image
If someone adds an image to a section it will be added to the sidebar
TOC of a PDF, too. A PDF viewer will show the "width and height"
data instead of an image. So we need to filter out images for this with
the optional "short title" of latex sections.