Fix #2298 automodule fails to document a class attribute.
This commit adds the method ClassDocumenter.generate. This method calls
the implementation from the super class without passing the optional 
argument real_modname. This causes Documenter.generate() to use the
module of the imported class. 

Why is this change save: if the class can't be imported in 
Documenter.generate(), Documenter.generate() returns early and the
module name is not used. If Documenter.generate() can import the class,
Documenter.get_real_modname() always returns a module name.