HTML escape item names in search results.
When the name of an item contains characters that need HTML escaping,
the search results need to escape the text so that it appears correctly.

Everything else is escaped correctly for item names apart from this.

This happens with the Dylan language where it is valid for identifiers
to contain characters such as '<' and '>'.

Fixes #2019.

* sphinx-themes/basic/static/searchtools.js_t
  (Search.dispalyNextItem): Use .text() rather than .html() so that
   the item name gets HTML escaped correctly.