writers.html.HTMLTranslator: Add space for wrap.
Adding a space after the `<code class="descname">...</code>` sections
and before the sig-paren span improves line wrapping significantly.
For example, a class method with a few parameters renders something
like the following by default:

    module.sub.Class.method_name(first_param,another_parameter,
    yet_another_param=None, **kwargs)                        [source]
      Short description.

By adding a single space, we can improve this to:

    module.sub.Class.method_name
    (first_param, another_param, yet_another_param=None,
    **kwargs)                                                [source]
      Short description.

which is somewhat cleaner.  Add in some CSS for descsignature:

    module.sub.Class.method_name
        (first_param, another_param, yet_another_param=None,
	**kwargs)
      Short description.