Extend pdflatex config in sphinx.sty to subparagraphs
This addresses #2547. As the two Sphinx classes set secnumdepth to 2 by
default, the label for paragraphs with the numbering is still skipped by
default. But if user increases secnumdepth to 4 or higher it will show,
contrarily to earlier situation. Also formerly paragraph headings used
\small, which looks wrong because the body text did not use \small.

And subparagraphs were not at all configured with \titleformat, thus did
not use same font as other headings.