Remove all memory addresses from autodoc
Change `object_desription()` to remove all hexadecimal addresses, not only those at the end of the string.

The `repr()` of some objects is generated by taking the `repr()` of one object and then adding stuff to it. Therefore, memory addresses like `<object foo at 0xabcdef>` do not need to occur at the end of the string.