Sync shipped tabulary with upstream one.
The update incorporates
\changes{v0.10}{2014/06/21}
      {support \cs{cellcolor} see
	  http://tex.stackexchange.com/a/185851/1090}
from tabulary commented source.

Memo: Sphinx ships a custom tabulary to fix a footnote issue, in
relation with package footnote:

   sphinx.sty does \makesavenoteenv{tabulary} but this needs also
   a patch inside tabulary package

It would be better, as package footnote has some bugs, for Sphinx to
ship with package footnotehyper which fixes theses bugs and can easily
be extended to ensure the compatibility with tabulary. This would be
better than shipping a custom tabulary.