latex: use in sphinx.sty \SPX or \spx@ prefix systematically
Also, rename some environments to start with SPX.

some exceptions: the macros in sphinx.sty starting with \py@, or \DU, or
\PYG as well as \ifsphinxpdfoutput and \sphinxAtStartFootnote have been
left as is. Also the literal-block environment which is removed in
another PR is left as is.

these changes such as \strong -> \SPXstrong, \code -> \SPXcode,
\titleref -> \SPXtitleref (fixing a memoir class incompatibility) etc..
will unavoidably break projets which used ``.. raw:: latex`` directive
or ``'preamble'`` key to customize them. But in the long run, using a
prefix is the best way to ensure user can add custom packages or own
latex code.

The command find . -name '*.sty' -exec grep -l \\\\spx@ {} \; and
similarly with \\\\SPX have been executed in TeXLive 2015 and 2016 root
latex repertories to make sure no other packages use \SPX or \spx@ as
prefixes.