use 'sphinxpackageoptions' as argument to ``\sphinxsetup``
Indeed, LaTeX's handling of package options is not robust when options
contain LaTeX macros; thus we pass all the user customization to
``\sphinxsetup``, not as package options to "sphinx.sty".

Also, make parameters for long code line breaking (ref #2343, 1.4.2)
accessible as sub-keys of 'sphinxpackageoptions'. The space reserved for
the continuation character has been modified for coherence and
compatibility with use in footnotes. In fact it did not use the correct
font for the dimension computation. The documentation explains how to
recover former value.