Add possibility to specify a config file.
This commit addresses #3133 and enhances the -c command line option.
"-c" may now either specify a directory or point to a config file. In
case of a directory, it is assumed to contain a conf.py file.
The behavior of "-c" remains fully backwards compatible.

This commit contains the (small) functional changes in the command line
parser and changes in the documentation. Most occurrences of
:file:`conf.py` have been replaced by a link to the configuration file
documentation where it makes sense.