Correct Spanish translation of Release
In this case, release should be translated to "versión", not to "publicación", since it matches the context more accurately.

See that version is used in French, versió in Catalan, versão in Portuguese from Portugal and versiune in Romanian. In other romance languages like Italian, the word release is used untranslated.