doc: Add guide on setuptool/distutils integration
While this feature is somewhat documented in the API guide, there's
nothing about this option in the main user guide. Given the lack of
blogs and articles talking about this feature, along with the amount of
projects that are only using the same copy-paste '[build_sphinx]'
definition, I imagine this is an issue.

Make the feature a little more accessible by adding a section to the
main user guide that details (a) the fact the feature exists, (b) why
you'd want to use it, and (c) how you can use it.

Signed-off-by: Stephen Finucane <stephen@that.guru>