Default SPHINXBUILD to 'python -msphinx' instead of 'sphinx-build'.
Using `python -msphinx` as the default for SPHINXBUILD in the
Makefile/make.bat ensures that whatever virtual environment is currently
active (if any) gets picked up.  In the current situation, imaging that
we are working on project $foo which is also installed system-wide,
but we want to build docs for a venv-installed version; that sphinx
is installed system-wide but not in the current environment, and that
e.g. conf.py imports $foo (e.g. to retrieve a version string).  Then
sphinx-build will resolve to the globally installed version, and the
global (wrong) version of $foo will be imported as well.

With this fix, `make html` will do everything in the correct virtualenv
(if any).  This switch is similar in spirit to e.g. the deprecation of
the `pyvenv` script in Py3.6, in favor of `python -mvenv`.