Remove superfluous ``\long`` from text styling macro definitions
All LaTeX commands such as \textbf, \emph, \underline.. are "short",
thus there was no need of ``\long`` prefix. Regarding
``\sphinxoptional`` which was defined via ``\newcommand``, the ``\long``
is there for full backwards compatibility, but a priori the argument
will always be a "short" one (i.e. not containing empty line or ``\par``
token.)