Streamline obsolete color set-up in LaTeX style file
The file ``pdfcolor.tex`` included in tex distributions is for Plain
pdfTeX, not for LaTeX use. Loading it complicates task of supporting
multiple TeX engines. If needed all corresponding colour names are
better available from passing to ``color`` or ``xcolor`` the
``dvipsnames`` option. It is better to leave responsability of using
this option to Sphinx user, if needed. The macros for setting the color
defined by ``pdfcolor.tex`` were never documented to Sphinx user, and
they can not be used from reST apart from ``.. raw:: latex`` directive,
they are for Plain TeX without use of ``color`` package. Thus, this
commit drops loading this file and the complicated conditionals around
it (``\ifsphinxpdfoutput``), which are also used in the class files
``sphinxmanual.cls``, ``sphinxhowto.cls``.

Also, this fixes issue #2693.