Fix links to external option docs with intersphinx.
When using intersphinx doc sets, the `:option:` directive is unable to
locate external option definitions without explicitly specifying an
additional target in the form of "<progname>.<option>". This is because
the directive is attempting to use state not available over intersphinx
to locate the appropriate directive. It's even worse if attempting to
define a ".. program::" entry in the documentation trying to link to an
external option.

To fix this, the intersphinx code for locating definitions now attempts
to convert option references into the right format. It does this by
rebuilding the string to normalize the program name and the option,
taking care to retain the intersphinx doc set name and the current
program if one isn't specified in the option. This allows external
options to be correctly resolved over intersphinx without custom syntax
or workarounds, behaving like other standard references.

Closes #2822.