FIDDLING: Change node type
Attempted changing node type of Cmdoption ObjectDescription from
addnodes.desc_name to addnodes.literal_strong, but it doesn't
fix the problem with how it displays. No obvious problems in
documentation make, but some tests failed. The same number of
tests failed when run on `stable`, however, so it appears to be
truly harmless/ineffectual.