latex: add option to put literal block captions before or after
According to ``literalblockcappos`` setting in ``'sphinxsetup'``, (t)op
or (b)ottom, the caption will be typeset before or after the literal
block. If typeset above, its distance to frame is now identical to the
setting used for captions of tables (one half of the baseline).

Further keys ``verbatimcontinuedalign`` and ``verbatimcontinuesalign``
allow to horizontally align the continuation hints relative to the
literal block contents: either (l)eft, (c)enter, or (r)ight.