Fix #3851 and fix #3706 about box drawing characters for PDF output
Do not tex-escape them, but provide suitable definitions for pdflatex
only (and platex). Define the 4 box drawing characters needed by unix
tree command with charset utf8. Also define the U+2572 as it was
previously escaped for all engines.

No definitions for xelatex/lualatex as it is then only a matter of using
a font providing them (a priori, only the mono font needs definition, as
the characters will be used in literal blocks).