Fix #1135: Make 'sphinx-apidoc' an extension
This is an alternative approach to #1135. Rather than modifying the
'build_sphinx' setuptools extension, we allow 'sphinx-apidoc' to be run
as an extension.

This is very similar to what we do with 'sphinx-autogen', which can be
run automatically using the 'autosummary_generate' configuration option
instead.

We may wish to remove the 'sphinx-apidoc' binary in the future, but
that's a decision for another day.

Signed-off-by: Stephen Finucane <stephen@that.guru>