nature: Fix macOS Safari scrollbar color
macOS scrollbars overlap the contents, and the color is either black or white,
depending on the color of the body. Therefore having a black body causes the
scrollbar to appear white, and since the theme itself is light, the scrollbar
is barely visible. Fix this by setting body to have a white background-color,
matching the visible color coming from div.body.