sphinx-build: Configure outdir from 'conf.py'
This is part one of two of the "make 'sphinx-build' easier to call"
effort. In this part, we allow users to define the build directory in
the 'conf.py' file. This means a user no longer needs to pass the option
via the command line, though they can continue to do so if they wish.

We use the behavior of the setuptools extension, whereby the results are
output to a subfolder in the outdir named after the code of the builder.
For example, if using the 'html' builder with 'output_dir = "test"',
files would be written to 'test/html'.

Signed-off-by: Stephen Finucane <stephen@that.guru>