Add autosectionlabel_max_depth config option
This option defines maximum section depth that labels will be
generated for by the autosectionlabel extension.

This is useful when there are a lot of sections of the similar
structure, for example:

 Releases
 ========
 Release 1
 ---------
 Changes
 ^^^^^^^

 Date
 ^^^^

 ...

 Release N
 ---------
 Changes
 ^^^^^^^

 Date
 ^^^^

This way there'll be warnings about duplicate labels. Setting
autosectionlabel_max_depth allows to skip sections deeper than
'Releases' in the example above.

By default it's None, so things will not change and labels will
be generated for all sections unless configured otherwise.