Remove docs of ``-M latexpdfja`` sphinx-build option
Indeed, if latex_engine is not appropriately set, the ``make latexpdf``
then fails because the Makefile in latex build repertory is the one for
pdflatex, not platex+dvipdfmx.

Besides, if language is 'ja', then latexpdf automatically uses the
platex latex_engine, hence the platex/dvipdfmx pipeline, and latexpdfja
is not needed.

For now, only removing documentation.